Rename v2 to v-2 for consistency
